refactor: move cd and ls to navigation directory
