+++ src/commands/navigation/cd.js
@@ -1,7 +1,7 @@
 import fs from "node:fs/promises";
 import path from "node:path";
-import { printInputError } from "../messages/inputError.js";
-import { printOperationError } from "../messages/operationError.js";
+import { printInputError } from "../../messages/inputError.js";
+import { printOperationError } from "../../messages/operationError.js";
 
 export const cd = async (command) => {
   const pathArg = getValidArg(command);
